requirements.txt
# Core Data and Visualization
numpy<2.0.0
pandas
matplotlib
pyserial

# Computer Vision & Hardware
opencv-python
depthai==3.3.0
blobconverter

# AI Frameworks (Version locked for OAK-D compatibility)
tensorflow-intel==2.15.0
openvino==2024.0.0